# Conflicts:
#	.idea/CoolRequestHistoryPersistent.xml
#	Service-User/src/main/java/org/example/serviceuser/admin/controller/PostAdminController.java

+++ Service-User/src/main/java/org/example/serviceuser/admin/request/BatchDeleteRequest.java
@@ -14,5 +14,5 @@ import lombok.Data;
 import java.util.List;
 @Data
 public class BatchDeleteRequest {
-    private List<Integer> ids;  // 用户 ID 列表
+    private List<Long> ids;  // 用户 ID 列表
 }
